feat(capas): composable de acciones repetidas
