+++ src/componentes/capa/xyz/SisdaiCapaXyz.vue
@@ -1,8 +1,8 @@
 <script setup>
-import { inject, onMounted, onUnmounted, reactive, toRefs, watch } from 'vue'
+import { inject, reactive, toRefs, watch } from 'vue'
 
 import TileLayer from 'ol/layer/Tile'
-import { getRenderPixel } from 'ol/render'
+// import { getRenderPixel } from 'ol/render'
 import ImageTile from 'ol/source/ImageTile'
 import TileEventType from 'ol/source/TileEventType'
 
@@ -10,28 +10,19 @@ import { MAPA_INYECTADO } from './../../../utiles/identificadores'
 import MonitoreoCargaElementos, {
   TipoEstadoCarga,
 } from './../../../utiles/MonitoreoCargaElementos'
+import useCapa from './../useCapa'
 import eventos from './eventos'
 import _props from './props'
 
 const mapa = inject(MAPA_INYECTADO)
 const emits = defineEmits(Object.values(eventos))
 const props = defineProps(_props)
-const { fuente, opacidad, posicion, titulo, visible } = toRefs(props)
+const { fuente, titulo } = toRefs(props)
 
 const source = new ImageTile({ url: fuente.value, crossOrigin: 'anonymous' })
 
-const capa = new TileLayer({
-  id: props.id,
-  opacity: opacidad.value,
-  source,
-  titulo: titulo.value,
-  visible: visible.value,
-  zIndex: posicion.value,
-})
-
-watch(opacidad, nuevaOpacidad => capa.setOpacity(nuevaOpacidad))
-watch(posicion, nuevaPosicion => capa.setZIndex(nuevaPosicion))
-watch(visible, nuevaVisibilidad => capa.setVisible(nuevaVisibilidad))
+const capa = new TileLayer({ id: props.id, source, titulo: titulo.value })
+useCapa(capa, props)
 
 // // mapa.capas = { ...mapa.capas, [props.id]: TipoEstadoCarga.no }
 // mapa.capas[props.id] = TipoEstadoCarga.no
@@ -61,34 +52,31 @@ source.on(TileEventType.TILELOADERROR, () => {
   monitoreoCargaTeselas.error++
 })
 
-onMounted(() => mapa.addLayer(capa))
-onUnmounted(() => mapa.quitarCapa(props.id))
-
 // { type, target, inversePixelTransform, frameState, context }
-capa.on('prerender', event => {
-  const tamanioMapa = mapa.getSize()
-  const nuevoAnchoCapa = tamanioMapa[0] * (Number(mapa.dividir) / 100)
-  const tl = getRenderPixel(event, [nuevoAnchoCapa, 0])
-  const tr = getRenderPixel(event, [tamanioMapa[0], 0])
-  const bl = getRenderPixel(event, [nuevoAnchoCapa, tamanioMapa[1]])
-  const br = getRenderPixel(event, tamanioMapa)
-  // console.log(event.type, tamanioMapa, tl, tr, bl, br)
-  const { context } = event
+// capa.on('prerender', event => {
+//   const tamanioMapa = mapa.getSize()
+//   const nuevoAnchoCapa = tamanioMapa[0] * (Number(mapa.dividir) / 100)
+//   const tl = getRenderPixel(event, [nuevoAnchoCapa, 0])
+//   const tr = getRenderPixel(event, [tamanioMapa[0], 0])
+//   const bl = getRenderPixel(event, [nuevoAnchoCapa, tamanioMapa[1]])
+//   const br = getRenderPixel(event, tamanioMapa)
+//   // console.log(event.type, tamanioMapa, tl, tr, bl, br)
+//   const { context } = event
 
-  context.save()
-  context.beginPath()
-  context.moveTo(tl[0], tl[1])
-  context.lineTo(bl[0], bl[1])
-  context.lineTo(br[0], br[1])
-  context.lineTo(tr[0], tr[1])
-  context.closePath()
-  context.clip()
-})
-capa.on('postrender', ({ context }) => {
-  // console.log(type)
+//   context.save()
+//   context.beginPath()
+//   context.moveTo(tl[0], tl[1])
+//   context.lineTo(bl[0], bl[1])
+//   context.lineTo(br[0], br[1])
+//   context.lineTo(tr[0], tr[1])
+//   context.closePath()
+//   context.clip()
+// })
+// capa.on('postrender', ({ context }) => {
+//   // console.log(type)
 
-  context.restore()
-})
+//   context.restore()
+// })
 </script>
 
 <template>
